refactor: make `App` more extensible using `BaseApp` (#14197)
