+++ benchmark/packages/adapter/src/server.ts
@@ -1,11 +1,11 @@
 import * as fs from 'node:fs';
 import type { SSRManifest } from 'astro';
-import { App } from 'astro/app';
+import { AppPipeline, BaseApp } from 'astro/app';
 import { applyPolyfills } from 'astro/app/node';
 
 applyPolyfills();
 
-class MyApp extends App {
+class MyApp extends BaseApp {
 	#manifest: SSRManifest | undefined;
 	constructor(manifest: SSRManifest, streaming = false) {
 		super(manifest, streaming);
@@ -21,6 +21,14 @@ class MyApp extends App {
 		}
 
 		return super.render(request);
+	}
+
+	createPipeline(streaming: boolean) {
+		return AppPipeline.create({
+			logger: this.logger,
+			manifest: this.manifest,
+			streaming,
+		});
 	}
 }
