+++ packages/astro/src/core/app/node.ts
@@ -6,7 +6,8 @@ import { clientAddressSymbol } from '../constants.js';
 import { deserializeManifest } from './common.js';
 import { createOutgoingHttpHeaders } from './createOutgoingHttpHeaders.js';
 import type { RenderOptions } from './index.js';
-import { App } from './index.js';
+import { BaseApp } from './index.js';
+import { AppPipeline } from './pipeline.js';
 import type { NodeAppHeadersJson, SerializedSSRManifest, SSRManifest } from './types.js';
 
 export { apply as applyPolyfills } from '../polyfill.js';
@@ -19,7 +20,14 @@ interface NodeRequest extends IncomingMessage {
 	body?: unknown;
 }
 
-export class NodeApp extends App {
+export class NodeApp extends BaseApp {
+	createPipeline(streaming: boolean): AppPipeline {
+		return AppPipeline.create({
+			logger: this.logger,
+			manifest: this.manifest,
+			streaming,
+		});
+	}
 	headersMap: NodeAppHeadersJson | undefined = undefined;
 
 	public setHeadersMap(headers: NodeAppHeadersJson) {
