+++ packages/astro/src/core/app/pipeline.ts
@@ -2,39 +2,37 @@ import type { ComponentInstance } from '../../types/astro.js';
 import type { RewritePayload } from '../../types/public/common.js';
 import type { RouteData, SSRElement, SSRResult } from '../../types/public/internal.js';
 import { Pipeline, type TryRewriteResult } from '../base-pipeline.js';
-import type { SinglePageBuiltModule } from '../build/types.js';
-import { RedirectSinglePageBuiltModule } from '../redirects/component.js';
-import { createModuleScriptElement, createStylesheetElementSet } from '../render/ssr-element.js';
+import {
+	createAssetLink,
+	createModuleScriptElement,
+	createStylesheetElementSet,
+} from '../render/ssr-element.js';
 import { findRouteToRewrite } from '../routing/rewrite.js';
 
 export class AppPipeline extends Pipeline {
 	static create({
 		logger,
 		manifest,
-		runtimeMode,
-		renderers,
-		resolve,
-		serverLike,
 		streaming,
-		defaultRoutes,
-	}: Pick<
-		AppPipeline,
-		| 'logger'
-		| 'manifest'
-		| 'runtimeMode'
-		| 'renderers'
-		| 'resolve'
-		| 'serverLike'
-		| 'streaming'
-		| 'defaultRoutes'
-	>) {
+	}: Pick<AppPipeline, 'logger' | 'manifest' | 'streaming'>) {
+		const resolve = async function resolve(specifier: string) {
+			if (!(specifier in manifest.entryModules)) {
+				throw new Error(`Unable to resolve [${specifier}]`);
+			}
+			const bundlePath = manifest.entryModules[specifier];
+			if (bundlePath.startsWith('data:') || bundlePath.length === 0) {
+				return bundlePath;
+			} else {
+				return createAssetLink(bundlePath, manifest.base, manifest.assetsPrefix);
+			}
+		};
 		const pipeline = new AppPipeline(
 			logger,
 			manifest,
-			runtimeMode,
-			renderers,
+			'production',
+			manifest.renderers,
 			resolve,
-			serverLike,
+			true,
 			streaming,
 			undefined,
 			undefined,
@@ -44,7 +42,6 @@ export class AppPipeline extends Pipeline {
 			undefined,
 			undefined,
 			undefined,
-			defaultRoutes,
 		);
 		return pipeline;
 	}
@@ -91,35 +88,5 @@ export class AppPipeline extends Pipeline {
 
 		const componentInstance = await this.getComponentByRoute(routeData);
 		return { newUrl, pathname, componentInstance, routeData };
-	}
-
-	async getModuleForRoute(route: RouteData): Promise<SinglePageBuiltModule> {
-		for (const defaultRoute of this.defaultRoutes) {
-			if (route.component === defaultRoute.component) {
-				return {
-					page: () => Promise.resolve(defaultRoute.instance),
-					renderers: [],
-				};
-			}
-		}
-
-		if (route.type === 'redirect') {
-			return RedirectSinglePageBuiltModule;
-		} else {
-			if (this.manifest.pageMap) {
-				const importComponentInstance = this.manifest.pageMap.get(route.component);
-				if (!importComponentInstance) {
-					throw new Error(
-						`Unexpectedly unable to find a component instance for route ${route.route}`,
-					);
-				}
-				return await importComponentInstance();
-			} else if (this.manifest.pageModule) {
-				return this.manifest.pageModule;
-			}
-			throw new Error(
-				"Astro couldn't find the correct page to render, probably because it wasn't correctly mapped for SSR usage. This is an internal error, please file an issue.",
-			);
-		}
 	}
 }
